Printing the destination register as W or F instead of 0 or 1

+++ trunk/pigmeo-compiler/src/BackendPIC/AsmInstruction.cs
@@ -187,7 +187,7 @@ namespace Pigmeo.Compiler.BackendPIC {
 					returned += OP.ToString() + " " + file;
 					break;
 				case InstructionType.ByteOriented_fd:
-					returned += OP.ToString() + " " + file + "," + (byte)DestinationWF;
+					returned += OP.ToString() + " " + file + "," + DestinationWF;
 					break;
 				case InstructionType.Literal_address:
 					returned += OP.ToString() + " " + LiteralAddr;
